requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
celery==5.3.1
redis==4.6.0
requests==2.31.0
python-dotenv==1.0.0
sentence-transformers
faiss-cpu
PyPDF2==3.0.1
numpy
openai